% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{AMINOACIDS}
\alias{AMINOACIDS}
\title{Symbol translation table for the 20 standard proteogenic amino acids}
\format{A tibble with 20 rows and 3 columns:
\describe{
  \item{full}{Full-name symbol}
  \item{three}{Three-letter symbol}
  \item{one}{One-letter symbol}
}}
\usage{
AMINOACIDS
}
\description{
Symbol translation table for the 20 standard proteogenic amino acids
}
\keyword{datasets}
