% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anlz_attainpep.R
\name{anlz_attainpep}
\alias{anlz_attainpep}
\title{Get attainment categories}
\usage{
anlz_attainpep(meddat, magdurout = FALSE, trgs = NULL)
}
\arguments{
\item{meddat}{result returned from \code{\link{anlz_medpep}}}

\item{magdurout}{logical indicating if the separate magnitude and duration estimates are returned}

\item{trgs}{optional \code{data.frame} for annual bay segment water quality targets, defaults to \code{\link{peptargets}}}
}
\value{
A \code{data.frame} for each year and bay segment showing the attainment category
}
\description{
Get attainment categories for each year and bay segment using chlorophyll and secchi depth
}
\examples{
meddat <- anlz_medpep(rawdat)
anlz_attainpep(meddat)
}
\seealso{
Other analyze: \code{\link{anlz_medpep}}
}
\concept{analyze}
