% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anlz_avepep.R
\name{anlz_avepep}
\alias{anlz_avepep}
\title{Estimate annual means}
\usage{
anlz_avepep(dat)
}
\arguments{
\item{dat\code{data.frame}}{formatted from \code{read_pepwq}}
}
\value{
Mean estimates for chlorophyll and secchi
}
\description{
Estimate annual means by segment for chlorophyll and secchi data
}
\examples{
# view average estimates
anlz_avepep(rawdat)
}
\concept{analyze}
