% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anlz_medpep.R
\name{anlz_medpep}
\alias{anlz_medpep}
\title{Estimate annual medians}
\usage{
anlz_medpep(dat)
}
\arguments{
\item{dat}{\code{data.frame} formatted from \code{\link{read_pepwq}}}
}
\value{
Median estimates for chlorophyll, secchi, and total nitrogen
}
\description{
Estimate annual medians by segment for chlorophyll, secchi, and total nitrogen data
}
\examples{
# view median estimates
anlz_medpep(rawdat)
}
\concept{analyze}
