% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beaches.R
\docType{data}
\name{beaches}
\alias{beaches}
\title{Bathing beaches}
\format{
A \code{data.frame} object
}
\usage{
beaches
}
\description{
Bathing beaches
}
\examples{
\dontrun{

library(sf)
library(dplyr)

beaches <- read.csv('inst/extdata/beaches.csv') \%>\% 
  st_as_sf(coords = c('Longitude', 'Latitude'), crs = '+proj=longlat +ellps=WGS84 +datum=WGS84 +no_defs')

names(beaches)[1] <- 'Name'

save(beaches, file = 'data/beaches.RData')

}
}
\seealso{
Other utilities: 
\code{\link{dodat}},
\code{\link{entdat}},
\code{\link{pepseg}},
\code{\link{pepstations}},
\code{\link{peptargets}},
\code{\link{rawdat}}
}
\concept{utilities}
\keyword{datasets}
