% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{BGFREQS}
\alias{BGFREQS}
\title{Amino Acid Residue Background Frequencies for Kongdom All}
\format{A named numeric vector with 20 elements:
\describe{
  \item{name}{One-letter symbol for each of the 20 standard proteogenic amino acids}
  \item{value}{The background frequency for kingdom all}
}}
\source{
\url{https://academic.oup.com/nar/article/45/D1/D1112/2333931}
}
\usage{
BGFREQS
}
\description{
Background distribution frequencies for Kullback-Leibler calculation is from:
    Proteome-pI: proteome isoelectric point database
    Lukasz P. Kozlowski
    Nucleic Acids Research, Volume 45, Issue D1, 4 January 2017, Pages D1112–
    D1116, https://doi.org/10.1093/nar/gkw978
}
\keyword{datasets}
