% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{BLOSUM50_enc}
\alias{BLOSUM50_enc}
\title{The BLOSUM50 matrix}
\format{
A data frame with with 20 rows and 20 columns
}
\source{
\url{ftp://ftp.ncbi.nih.gov/blast/matrices/BLOSUM50}
}
\usage{
BLOSUM50_enc
}
\description{
The BLOSUM50 matrix transformed for encoding by dividing by 5. Furthermore,
row and column zero vectors representing 'X' has been added.
}
\details{
The BLock SUbstition matrix for sequences with less than 50\\% similarity.
The matrix has been subset to the 20 proteogenic amino acids

Matrix made by matblas from blosum50.iij
\itemize{
\item column uses minimum score
}

BLOSUM Clustered Scoring Matrix in 1/3 Bit Units

Blocks Database = /data/blocks_5.0/blocks.dat

Cluster Percentage: >= 50

Entropy =   0.4808, Expected =  -0.3573
}
\keyword{datasets}
