% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{BLOSUM50_pca}
\alias{BLOSUM50_pca}
\title{The BLOSUM50 encoding pca matrix}
\format{
A data frame with with 21 rows and 20 columns
}
\source{
\url{ftp://ftp.ncbi.nih.gov/blast/matrices/BLOSUM50}
}
\usage{
BLOSUM50_pca
}
\description{
The BLOSUM50 matrix transformed for encoding by performing a PCA
and then using n number of PCs for encoding
}
\keyword{datasets}
