% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{BLOSUM62}
\alias{BLOSUM62}
\title{The BLOSUM62 matrix}
\format{A named numeric matrix with 20 rows and 20 columns}
\source{
\url{https://www.ncbi.nlm.nih.gov/Class/FieldGuide/BLOSUM62.txt}
}
\usage{
BLOSUM62
}
\description{
The BLock SUbstition matrix for sequences with less than 62\% similarity
}
\keyword{datasets}
