% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{BLOSUM62_FREQ}
\alias{BLOSUM62_FREQ}
\title{The row normalised BLOSUM62 frequency matrix}
\format{A named numeric matrix with 20 rows and 20 columns}
\source{
\url{http://www.cs.cornell.edu/courses/cs626/2003sp/blosum62qij.txt}
}
\usage{
BLOSUM62_FREQ
}
\description{
The BLock SUbstition matrix for sequences with less than 62\% similarity
}
\keyword{datasets}
