% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{BLOSUM62_pca}
\alias{BLOSUM62_pca}
\title{The BLOSUM62 encoding pca matrix}
\format{
A data frame with with 21 rows and 20 columns
}
\source{
\url{ftp://ftp.ncbi.nih.gov/blast/matrices/BLOSUM62}
}
\usage{
BLOSUM62_pca
}
\description{
The BLOSUM62 matrix transformed for encoding by performing a PCA
and then using n number of PCs for encoding
}
\keyword{datasets}
