% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{BLOSUM62_PROB}
\alias{BLOSUM62_PROB}
\title{The row normalised BLOSUM62* frequency matrix}
\format{A named numeric matrix with 20 rows and 20 columns}
\source{
\url{http://www.cs.cornell.edu/courses/cs626/2003sp/blosum62qij.txt}
}
\usage{
BLOSUM62_PROB
}
\description{
*The BLock SUbstition matrix for sequences with less than 62\% similarity
}
\details{
Rows are normalised such that sum(X_i) = 1 for all i in X (BLOSUM62)
}
\keyword{datasets}
