% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_install_TMB.R
\name{check_install_TMB}
\alias{check_install_TMB}
\title{Check TMB is working}
\usage{
check_install_TMB()
}
\value{
A single integer is returned, where zero means that your
installation of rtools and TMB is good to go. I hope you get a zero! If the
status is failed, then you will see \code{"Compilation failed"}, a stop message
generated by \code{\link[TMB:compile]{TMB::compile()}}.
}
\description{
Check that the installation of TMB is working before trying to run any TMB
models.
}
\details{
This function will generate a lot of output to your screen that can be
helpful when trying to troubleshoot the installation of TMB. Most
importantly, you can see the location of where your computer is trying to
find make files. If you have already ran this function on your computer,
then you should see \verb{make: Nothing to be done for 'all'}. If the
installation of TMB has failed, then you will see \code{"Error in TMB::compile(...)"}. The most likely culprit for it failing is a faulty
installation of rtools, specifically rtools40 for R version 4.0 and 4.1 and
rtools42 for R version 4.20 and greater.
}
\examples{
# Hope for a 0
\dontrun{
check_install_TMB()
}
# If you get something other than 0, run
pkgbuild::find_rtools()

}
\seealso{
\itemize{
\item \code{pkgbuild::find_rtools[]}
}
}
\author{
Kelli F. Johnson
}
