% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entdat.R
\docType{data}
\name{entdat}
\alias{entdat}
\title{Raw beach pathogen data from Suffolk County}
\format{
A \code{data.frame} object
}
\usage{
entdat
}
\description{
Raw beach pathogen data from Suffolk County
}
\examples{
\dontrun{
entdat <- read_pepent()
save(entdat, file = 'data/entdat.RData', compress = 'xz')
}
}
\concept{data}
\keyword{datasets}
