% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buffer_fxn.R
\name{get_buffer}
\alias{get_buffer}
\title{Buffer calculation for U.S. West Coast Groundfish Stocks
Applies the time-varying sigma approach based on the 2019
approach.}
\usage{
get_buffer(years, sigma, pstar)
}
\arguments{
\item{years}{the years to generate buffer values for}

\item{sigma}{the sigma value to use (cat 1 = 0.50, cat 2 = 1.0, cat 3 = 2.0)}

\item{pstar}{managment uncertainty value (i.e., 0.45, 0.40)}
}
\description{
Buffer calculation for U.S. West Coast Groundfish Stocks
Applies the time-varying sigma approach based on the 2019
approach.
}
\examples{
get_buffer(years = 2011:2032, sigma = 1.0, pstar = 0.4)
}
\author{
written by Chantel Wetzel
}
