% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_other.R
\name{mat_mima_scale}
\alias{mat_mima_scale}
\title{Min-max scale matrix}
\usage{
mat_mima_scale(x, na_rm = FALSE)
}
\arguments{
\item{x}{A numeric matrix}

\item{na_rm}{Boolean: Should NAs be removed}
}
\description{
Min-max scale matrix, such that all values v lie in [0;1]. If X is a numeric
matrix, then min-max scale is: ( X - min(X) ) / ( max(X) - min(X) )
}
\examples{
mat_mima_scale(matrix(rnorm(180),nrow=9,ncol=20))
}
