% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ONEHOT_enc}
\alias{ONEHOT_enc}
\title{The ONEHOT encoding matrix}
\format{
A data frame with with 21 rows and 21 columns
}
\usage{
ONEHOT_enc
}
\description{
Here, every amino acid residue is simply encoded orthogonal using a
diagonal unit matrix
}
\keyword{datasets}
