% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pacfin_ports_withlatlong}
\alias{pacfin_ports_withlatlong}
\title{Port Location Data}
\format{
\subsection{\code{pacfin_ports_withlatlong}}{

A data frame with 507 rows and 6 columns:
\describe{
\item{name, agencydesc}{U.S. West Coast port name.}
\item{pcid}{3 letter port code used in PacFIN}
\item{agid}{Letter indicating the location of port within a state,
where W = Washington, O = Oregon, and C = California.}
\item{latitude, longitude}{Latitude/Longitude of the port in radians.}
}
}
}
\usage{
pacfin_ports_withlatlong
}
\description{
Latitude and longitude of West Coast ports.
}
\keyword{datasets}
