% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_pep.R
\name{pep_check}
\alias{pep_check}
\title{Check if a vector of peptides is valid.}
\usage{
pep_check(pep)
}
\arguments{
\item{pep}{A vector of peptides to be checked}
}
\value{
TRUE if the vector passed the check, otherwise stop() is called
}
\description{
According to criteria:
 i. Is a vector with one or more elements.
 ii. Is a character vector.
 iii. All elements in vector have the same number of characters.
 iv. Elements only contain allowed amino acid residues \code{'ARNDCQEGHILKMFPSTWYVX-'}.
}
\examples{
pep_check(c("RQGQDHPTM","RGQKTTDNA","NILYEYWDY"))
pep_check(c("RQGQDHPTM","RGQKTTDNA","NILYEYWDYX"))
pep_check(c("RQGQDHPTM","RGQKTTDNA","NILYEYWDB"))
}
