% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_pep.R
\name{pep_clean}
\alias{pep_clean}
\title{Translate all non-standard amino acid residue characters with 'X'}
\usage{
pep_clean(pep)
}
\arguments{
\item{pep}{A character vector of peptides to be cleaned}
}
\value{
A character vector of cleaned peptides
}
\description{
Translate all non-standard amino acid residue characters with 'X'
}
\examples{
pep_clean(sample(letters,20,replace=TRUE))
}
