% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_pep.R
\name{pep_consensus}
\alias{pep_consensus}
\title{Retrieve the consensus sequence from a set of peptides}
\usage{
pep_consensus(pep)
}
\arguments{
\item{pep}{A vector of peptides}
}
\value{
A string corresponding to the consensus sequence
}
\description{
Retrieve the consensus sequence from a set of peptides
}
\examples{
pep_consensus(c("YMNSMQEML","FIYRHMFCV","VLFKFDMFI","KLLDRFPVA","RVLDDFTKL"))
}
