% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_pep.R
\name{pep_encode}
\alias{pep_encode}
\title{Encode a peptide}
\usage{
pep_encode(pep)
}
\arguments{
\item{pep}{A character vector of peptides to be encoded}
}
\value{
A 3D array (tensor) of peptide 'images'
}
\description{
Peptides are encoded using the BLOSUM62 matrix scaled to [0;1]
}
\details{
Each position in the peptide become a vector of 20 values, corresponding to
the rounded log odds ratio for substituting the amino acid in the peptide
with each of the 20 standard proteogenic amino acids.

The final result is a 3D array (tensor) of peptide 'images' with \code{n} rows, \code{k}
columns and \code{20} slices, which is stored as a 3D array (tensor)
}
\examples{
pep_encode(pep_ran(k=9,n=10))
dim(pep_encode(pep_ran(k=9,n=10)))
}
