% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_pep.R
\name{pep_encode_mat}
\alias{pep_encode_mat}
\title{Encode a peptide}
\usage{
pep_encode_mat(pep)
}
\arguments{
\item{pep}{A character vector of peptides to be encoded}
}
\value{
A tibble of encoded peptides
}
\description{
Peptides are encoded using the BLOSUM62 matrix scaled to [0;1]
}
\details{
Each position in the peptide become a vector of 20 values, corresponding to
the rounded log odds ratio for substituting the amino acid in the peptide
with each of the 20 standard proteogenic amino acids.

The final result is a tibble with as many rows as input peptides and number
columns equal to the number of positions in the input peptides times the
number of BLOSUM62 substitions values (20) plus a column with the input
peptides. For 100 9-mers this yields 100 rows and 181 columns
}
\examples{
pep_encode_mat(pep_ran(k=9,n=100))
dim(pep_encode_mat(pep_ran(k=9,n=100)))
}
