% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_pep.R
\name{pep_get_images}
\alias{pep_get_images}
\title{Get peptide images}
\usage{
pep_get_images(pep)
}
\arguments{
\item{pep}{A character vector of peptides to be converted to 'images'}
}
\value{
A list of peptide 'images'
}
\description{
Peptides are encoded using \code{pep_encode()}, resulting in peptide 'images'
}
\details{
Each position in the peptide become a vector of 20 values, corresponding to
the rounded log odds ratio for substituting the amino acid in the peptide
with each of the 20 standard proteogenic amino acids. The final result is a
list of peptide 'images'
}
\examples{
pep_get_images(pep_ran(k=9,n=10))
}
