% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_pep.R
\name{pep_mat}
\alias{pep_mat}
\title{Convert a vector of peptides to matrix form}
\usage{
pep_mat(pep)
}
\arguments{
\item{pep}{A vector of equal length peptides}
}
\value{
A n_pep x l_pep matrix
}
\description{
Convert a vector of peptides to matrix form
}
\examples{
pep_mat(c("RQGQDHPTM","RGQKTTDNA","NILYEYWDY"))
}
