% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_pep.R
\name{pep_mutate}
\alias{pep_mutate}
\title{Mutate residues along peptide positions}
\usage{
pep_mutate(pep, pos, res, keep_wt = FALSE)
}
\arguments{
\item{pep}{A string representing a peptide}

\item{pos}{Which position(s) to mutate}

\item{res}{Which amino acid(s) to mutate to}

\item{keep_wildtype}{Boolean should \code{pep} be included in the output}
}
\value{
A character vector with number of elements corresponding to the
number of peptides required to cover all mutations
}
\description{
Mutate residues along peptide positions
}
\examples{
pep_mutate(pep = "RQGQDHPTM", pos = seq(1,9), res = AMINOACIDS$one)
}
