% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_pep.R
\name{pep_plot_images}
\alias{pep_plot_images}
\title{Plot peptide images}
\usage{
pep_plot_images(pep, n = 3)
}
\arguments{
\item{pep}{A character vector of peptides to be plotted as 'images'}
}
\value{
A cowplot plot grid of the \code{n} first peptide 'images'
}
\description{
Plot the first \code{n} encoded peptide 'images'
}
\details{
Peptides are encoded using \code{pep_encode()}, resulting in peptide 'images'
Each position in the peptide become a vector of 20 values, corresponding to
the rounded log odds ratio for substituting the amino acid in the peptide
with each of the 20 standard proteogenic amino acids. The final result is a
list of peptide 'images' each with values in \code{[0;1]}.
}
\examples{
pep_plot_images(pep_ran(k=9,n=10))
}
