% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_pep.R
\name{pep_split}
\alias{pep_split}
\title{Split a peptide into residue characters}
\usage{
pep_split(pep)
}
\arguments{
\item{pep}{A string representing a peptide}
}
\value{
A character vector with number of elements corresponding to the number of amino acids in the input peptide
}
\description{
Split a peptide into residue characters
}
\examples{
pep_split("RQGQDHPTM")
}
