% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PEPmeeting.R
\name{PEPmeeting}
\alias{PEPmeeting}
\title{Generate Meeting Lead and Snack People}
\usage{
PEPmeeting(
  date,
  notattending = NA,
  leads = c("Aaron", "Andi", "Brian", "Chantel", "Jason", "John", "Kelli", "Kristin",
    "Ian", "Melissa", "Vlada"),
  snacks = c("Kathryn")
)
}
\arguments{
\item{date}{A numeric value giving the date of the meeting.}

\item{notattending}{A vector of names of those not attending.}

\item{leads}{A vector of potential leads.}

\item{snacks}{A vector of people that can bring snacks but
are not leads.}
}
\description{
Generate Meeting Lead and Snack People
}
\examples{
PEPtools::PEPmeeting(20200916, notattending=c("Owen", "Melissa"))
}
