% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pepseg.R
\docType{data}
\name{pepseg}
\alias{pepseg}
\title{Polygon shapefile of segment boundaries}
\format{
A \code{sf} object
}
\usage{
pepseg
}
\description{
Polygon shapefile of segment boundaries
}
\examples{
\dontrun{
library(sf)
library(dplyr)

pepseg <- st_read('~/Desktop/TBEP/Peconic/PEPSegments/PEP_Segments.shp') \%>\%
  st_transform(crs = '+proj=longlat +ellps=WGS84 +datum=WGS84 +no_defs') \%>\% 
  mutate(
    bay_segment = case_when(
      NAME == 'Segment3' ~ 'Eastern', 
      NAME == 'Segment2' ~ 'Central', 
      NAME == 'Segment1' ~ 'Western', 
    )
  ) \%>\% 
  select(bay_segment) \%>\% 
  st_buffer(dist = 0) \%>\% 
  st_crop(xmin = -72.76, xmax = -71.8, ymin = 40.83, ymax = 41.2)
 
save(pepseg, file = 'data/pepseg.RData', compress = 'xz')
}
}
\seealso{
Other utilities: 
\code{\link{dodat}},
\code{\link{entdat}},
\code{\link{pepstations}},
\code{\link{peptargets}},
\code{\link{rawdat}}
}
\concept{utilities}
\keyword{datasets}
