% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{PEPTIDES}
\alias{PEPTIDES}
\title{5,000 9-mer peptides}
\format{A character vector with 5,000 strings of length 9}
\usage{
PEPTIDES
}
\description{
5,000 9-mer peptides predicted to be strong binders to HLA-A*02:01 by netMHCpan-4.0
}
\keyword{datasets}
