% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pfmc_starter.R
\name{pfmc_starter}
\alias{pfmc_starter}
\title{Write a default starter.ss file for PFMC assessments}
\usage{
pfmc_starter()
}
\value{
A \code{starter.ss} file is saved to the current working directory.
}
\description{
Write a \code{starter.ss} file for assessments presented to the
Pacific Fisheries Management Council (PFMC) using a \code{starter.ss_new} file
from the Stock Synthesis GitHub organization with additional comments.
}
\author{
Vladlena Gertseva, Kelli F. Johnson, Brian J. Langseth,
Kristin Marshall, Kiva L. Oken, Ian G. Taylor, John R. Wallace, and
Sophia Wasserman
}
