% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_pssm.R
\name{pssm_check}
\alias{pssm_check}
\title{Check if a PSSM has the correct format}
\usage{
pssm_check(PSSM)
}
\arguments{
\item{PSSM}{A Position Specific Scoring Matrix}
}
\value{
\code{TRUE} if checks are passed, otherwise \code{stop()} is called with a message
}
\description{
Check format according to criteria: i. Is a matrix, ii. Is numeric,
iii. Has 20, 21 or 22 columns, iv. Column names only contain allowed
amino acid residues '\code{ARNDCQEGHILKMFPSTWYV}'
}
\examples{
pssm = pssm_empty()
pssm[1:9,1:20] = rnorm(180)
pssm_check(PSSM = pssm)
}
