% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_pssm.R
\name{pssm_empty_long}
\alias{pssm_empty_long}
\title{Create empty Position Specific Scoring Matrix}
\usage{
pssm_empty_long(npos = 9)
}
\arguments{
\item{npos}{The number of positions}
}
\value{
A numeric matrix with '0' entries with dimensions \code{npos} rows
and 22 columns
}
\description{
Create an empty Position Specific Scoring Matrix with \code{npos} number of
rows and 22 columns sorted according to \code{ARNDCQEGHILKMFPSTWYVX-}
}
\examples{
pssm_empty_long()
}
