% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_pssm.R
\name{pssm_flatten}
\alias{pssm_flatten}
\title{'Flatten' PSSM to a single row}
\usage{
pssm_flatten(PSSM)
}
\arguments{
\item{PSSM}{A Position Specific Scoring Matrix with \code{m}}
}
\value{
A numeric matrix with 1 row and m * n columns
}
\description{
'Flatten' PSSM to a single row, such that e.g. m x n = 9 x 20 becomes 1 x 180
}
\examples{
pssm_empty() \%>\% pssm_flatten
}
