% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_pssm.R
\name{pssm_freqs}
\alias{pssm_freqs}
\title{Compute frequency matrix}
\usage{
pssm_freqs(pep)
}
\arguments{
\item{pep}{A character vector of equal length peptides}
}
\value{
A numeric matrix with \code{length(pep)} rows and 20 columns
}
\description{
From a vector of equal length peptides, calculate the corresponding frequency matrix
}
\examples{
pep_ran(n = 100, k = 9) \%>\% pssm_freqs
}
