% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_pssm.R
\name{pssm_kld}
\alias{pssm_kld}
\title{Compute the Kullback-Leibler Divergence}
\usage{
pssm_kld(PSSM)
}
\arguments{
\item{PSSM}{A Position Specific Scoring Matrix of amino acid frequencies}
}
\value{
A PSSM of computed Kullback-Leibler Divergences for each \code{[p_i,res_j]}
}
\description{
From a matrix of amino acid frequencies, where the peptide positions are the
rows and where the amino acids are the columns and the sum of the
rows all are equal to one, calculate the compute the Kullback-Leibler
Divergence for each \code{[p_i,res_j]}
}
\examples{
pep_ran(n = 100, k = 9) \%>\% pssm_freqs \%>\% pssm_kld
}
