% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_install.R
\name{r_install}
\alias{r_install}
\title{Install packages for R}
\usage{
r_install()
}
\description{
Install devtools with all of its dependencies, i.e.,
\code{dependencies = c("Depends", "Imports", "LinkingTo", "Suggests")}
such that many packages that are needed downstream are installed.
}
\details{
Install packages from the cloud R-project repository using a single
call to install \pkg{devtools}. A single call is facilitated by
including \code{"Suggests"} in the dependencies argument.
The functionality of \code{r_install} is mainly useful for those
that update R on a regular basis. For example, developers of R packages
will want to install the development version of R at least monthly.
}
\examples{
\dontrun{
# The following will take at least 15 minutes to run on your machine
# if you only have the base R packages.
r_install()
}

}
\author{
Kelli Faye Johnson
}
