% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_other.R
\name{read_fasta}
\alias{read_fasta}
\title{Read a FASTA file}
\usage{
read_fasta(file)
}
\arguments{
\item{file}{The FASTA file to be read either local or URL}
}
\description{
read_fasta reads a FASTA file of nucleotides or amino acids file and returns
a tibble with number of rows corresponding to the number of sequences and two
variables: 'fasta_header' and 'sequences'
}
\examples{
read_fasta(file = 'my_fasta_file.fsa')
read_fasta(file = 'https://www.ncbi.nlm.nih.gov/WebSub/html/help/sample_files/nucleotide-sample.txt')
}
