% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_pepdo.R
\name{read_pepdo}
\alias{read_pepdo}
\title{Import dissolved oxygen data}
\usage{
read_pepdo(
  site = c("01304562", "01304200", "01304650"),
  nms = c("Peconic River", "Orient Harbor", "Shelter Island"),
  startDate = "2013-01-01",
  endDate = "2022-12-31"
)
}
\arguments{
\item{site}{chr string of site numbers for USGS stations}

\item{nms}{chr string vector of names to reassign to site numbers}

\item{startDate}{chr string of start date in YYYY-MM-DD}

\item{endDate}{chr string of end date in YYYY-MM-DD}
}
\value{
data.frame
}
\description{
Import dissolved oxygen data
}
\details{
Raw data are downloaded using the USGS dataRetrieval R package, this function is a simple wrapper to the \code{\link[dataRetrieval]{readNWISuv}} function.

Note that downloading the station data with the default arguments may take a few minutes. Sites are 01304562 for Peconic River, 01304200 for Orient Harbor, 01304650 for Shelter Island.
}
\examples{
\dontrun{
dodat <- read_pepdo(site = c('01304562', '01304200', '01304650'), 
  nms = c('Peconic River', 'Orient Harbor', 'Shelter Island'), 
  StartDate = '2020-06-01', endDate = '2021-06-30')
dodat
}
}
\concept{read}
