% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_pepwq.R
\name{read_pepwq}
\alias{read_pepwq}
\title{Import raw data}
\usage{
read_pepwq(path)
}
\arguments{
\item{path}{chr string of path for excel file}
}
\value{
data.frame
}
\description{
Impot raw data
}
\details{
Raw data from here \url{https://gisportal.suffolkcountyny.gov/gis/home/item.html?id=8107f192ffac406380b6d61d3d3dbf7d}
}
\examples{
path <- system.file("extdata", "currentdata.xlsx", package="pepreporting")
dat <- read_pepwq(path)
dat
}
\concept{read}
