% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_pep.R
\name{residue_name}
\alias{residue_name}
\title{Translate between amino acid residue symbols}
\usage{
residue_name(res, to = "one")
}
\arguments{
\item{res}{A character vector of residue symbols either in one-, three- or
full format}

\item{to}{Which symbol to translate to, one of \code{'one', 'three', 'full'}}
}
\value{
A character vector with the translated symbols
}
\description{
Translate between amino acid residue symbols
}
\examples{
residue_name(AMINOACIDS$full, to = 'three')
}
