% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_regularization.r
\name{run_regularization}
\alias{run_regularization}
\title{Run MCMC using {adnuts} for SS3}
\usage{
run_regularization(
  dir_wd,
  model = "ss3",
  extension = ".exe",
  interactive = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{dir_wd}{A file path to the directory where the model files and the
executable are saved. This can be relative to your current working
directory or a full path. Your working directory will be changed while
executing the function to be inside the location provided in this
argument but upon exiting it will return to your current working
directory.}

\item{model}{A string giving the name of the executable without an
an extension. The extension is provided in \code{extension}. Do not include
any path information, just the name of the executable. For example, the
default is \code{"ss3"}.}

\item{extension}{A string giving the extension for the executable name
provided in \code{model}. The default is \code{.exe}, which works on Windows
machines.}

\item{interactive}{A logical, where \code{TRUE} will run
\code{\link[adnuts:launch_shinyadmb]{adnuts::launch_shinyadmb()}}. The default is \code{FALSE}.}

\item{verbose}{A logical, specifying if information should be printed
to the screen. The default is \code{FALSE} but using \code{TRUE} can be helpful
when you are first running the function or when you have a very unstable
model.}
}
\description{
Run MCMC using {adnuts} for SS3
}
\author{
Cole M. Monnahan (cole.monnahan@noaa.gov; February 2021) and
augmented by Kelli F. Johnson
}
