% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_allthrpep.R
\name{show_allthrpep}
\alias{show_allthrpep}
\title{Plot annual water quality value and thresholds for all segments}
\usage{
show_allthrpep(
  dat,
  param = c("chla", "sd", "tn"),
  trgs = NULL,
  yrrng = NULL,
  family = NA,
  labelexp = TRUE,
  txtlab = TRUE
)
}
\arguments{
\item{dat}{data frame of data returned by \code{\link{read_pepwq}}}

\item{param}{chr string indicating which water quality value and appropriate threshold to plot, one of "chla" for chlorophyll, "sd" for secchi depth, or "tn" for total nitrogen}

\item{trgs}{optional \code{data.frame} for annual bay segment water quality thresholds, defaults to \code{\link{peptargets}}, only applies if \code{param} is \code{"chla"} or \code{"sd"}}

\item{yrrng}{numeric vector indicating min, max years to include}

\item{family}{optional chr string indicating font family for text labels}

\item{labelexp}{logical indicating if y axis and target labels are plotted as expressions, default \code{TRUE}}

\item{txtlab}{logical indicating if a text label for the target value is shown in the plot, only applies if \code{param} is \code{"chla"} or \code{"sd"}}
}
\value{
A \code{\link[ggplot2]{ggplot}} object
}
\description{
Plot annual water quality values and thresholds for all bay segments
}
\details{
This function is conceptually similar to \code{\link{show_thrpep}}, but results are shown as annual medians across all bay segments for the selected parameter.
}
\examples{
show_allthrpep(rawdat, param = 'chl')
}
\concept{visualize}
