% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_boxpep.R
\name{show_boxpep}
\alias{show_boxpep}
\title{Plot monthly chlorophyll, secchi, or tn values values for a segment}
\usage{
show_boxpep(
  dat,
  param = c("chla", "sd", "tn"),
  yrsel = NULL,
  yrrng = NULL,
  ptsz = 0.5,
  bay_segment = c("1a", "1b", "2", "3"),
  trgs = NULL,
  family = NA,
  labelexp = TRUE,
  txtlab = TRUE
)
}
\arguments{
\item{dat}{data frame of data returned by \code{\link{read_pepwq}}}

\item{param}{chr string indicating which water quality value and appropriate threshold to plot, one of "chla" for chlorophyll, "sd" for secchi depth, or "tn" for total nitrogen}

\item{yrsel}{numeric for year to emphasize, shown as separate red points on the plot}

\item{yrrng}{numeric vector indicating min, max years to include}

\item{ptsz}{numeric indicating point size of observations not in \code{yrsel}}

\item{bay_segment}{chr string for the bay segment, one of "1a", "1b", "2", or "3"}

\item{trgs}{optional \code{data.frame} for annual bay segment water quality targets, defaults to \code{\link{peptargets}}}

\item{family}{optional chr string indicating font family for text labels}

\item{labelexp}{logical indicating if y axis and target labels are plotted as expressions, default \code{TRUE}}

\item{txtlab}{logical indicating if a text label for the target value is shown in the plot}
}
\value{
A \code{\link[ggplot2]{ggplot}} object
}
\description{
Plot monthly chlorophyl, secchi, or tn values values for a bay segment
}
\details{
Points not included in \code{yrsel} are plotted over the box plots using \code{\link[ggplot2]{position_jitter}}. Use \code{ptsz = -1} to suppress.
}
\examples{
show_boxpep(rawdat, bay_segment = '1a')
}
\concept{visualize}
