% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_entmatrix.R
\name{show_entmatrix}
\alias{show_entmatrix}
\title{Create a colorized table for beach pathogen exceedances}
\usage{
show_entmatrix(entdat, txtsz = 2, thr = 104, yrrng = NULL, family = NA)
}
\arguments{
\item{entdat}{data frame of enterococcus data returned by \code{\link{read_pepent}}}

\item{txtsz}{numeric for size of text in the plot, applies only if \code{asreact = FALSE}}

\item{thr}{numeric value defining threshold for exceedance}

\item{yrrng}{numeric vector indicating min, max years to include}

\item{family}{optional chr string indicating font family for text labels}
}
\value{
A static \code{\link[ggplot2]{ggplot}} object is returned if \code{asreact = FALSE}, otherwise a \code{\link[reactable]{reactable}} table is returned
}
\description{
Create a colorized table for beach pathogen exceedances
}
\examples{
show_entmatrix(entdat)
}
\seealso{
\code{\link{anlz_entpep}}
}
\concept{visualize}
