% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_matrixpep.R
\name{show_matrixpep}
\alias{show_matrixpep}
\title{Create a colorized table for indicator reporting}
\usage{
show_matrixpep(
  dat,
  txtsz = 3,
  trgs = NULL,
  yrrng = NULL,
  bay_segment = c("Western", "Central", "Eastern"),
  asreact = FALSE,
  nrows = 10,
  abbrev = FALSE,
  family = NA
)
}
\arguments{
\item{dat}{data frame of water quality data returned by \code{\link{read_pepwq}}}

\item{txtsz}{numeric for size of text in the plot, applies only if \code{asreact = FALSE}}

\item{trgs}{optional \code{data.frame} for annual bay segment water quality targets, defaults to \code{\link{peptargets}}}

\item{yrrng}{numeric vector indicating min, max years to include}

\item{bay_segment}{chr string for bay segments to include, one to all of "Western", "Central", or "Eastern"}

\item{asreact}{logical indicating if a \code{\link[reactable]{reactable}} object is returned}

\item{nrows}{if \code{asreact = TRUE}, a numeric specifying number of rows in the table}

\item{abbrev}{logical indicating if text labels in the plot are abbreviated as the first letter}

\item{family}{optional chr string indicating font family for text labels}
}
\value{
A static \code{\link[ggplot2]{ggplot}} object is returned if \code{asreact = FALSE}, otherwise a \code{\link[reactable]{reactable}} table is returned
}
\description{
Create a colorized table for indicator reporting
}
\examples{
show_matrixpep(rawdat)
}
\seealso{
Other visualize: 
\code{\link{show_boxpep}()},
\code{\link{show_domatrix}()},
\code{\link{show_entmatrix}()},
\code{\link{show_plotlypep}()},
\code{\link{show_reactablepep}()},
\code{\link{show_segmatrixpep}()},
\code{\link{show_sitemappep}()},
\code{\link{show_thrpep}()},
\code{\link{show_wqmatrixpep}()}
}
\concept{visualize}
