% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_reactablepep.R
\name{show_reactablepep}
\alias{show_reactablepep}
\title{Create reactable table from matrix data}
\usage{
show_reactablepep(totab, colfun, nrows = 10)
}
\arguments{
\item{totab}{A data frame in wide format of summarized results}

\item{colfun}{Function specifying how colors are treated in cell background}

\item{nrows}{numeric specifying number of rows in the table}
}
\value{
A \code{\link[reactable]{reactable}} table
}
\description{
Create reactable table from matrix data
}
\details{
This function is used internally within \code{\link{show_matrixpep}}
}
\seealso{
Other visualize: \code{\link{show_boxpep}},
  \code{\link{show_matrixpep}},
  \code{\link{show_plotlypep}},
  \code{\link{show_sitemappep}}, \code{\link{show_thrpep}},
  \code{\link{show_wqmatrixpep}}
}
\concept{visualize}
