% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_segmatrixpep.R
\name{show_segmatrixpep}
\alias{show_segmatrixpep}
\title{Create a colorized table for water quality outcomes and exceedances by segment}
\usage{
show_segmatrixpep(
  dat,
  txtsz = 3,
  trgs = NULL,
  yrrng = NULL,
  bay_segment = c("1a", "1b", "2", "3"),
  abbrev = FALSE,
  family = NA
)
}
\arguments{
\item{dat}{\code{data.frame} formatted from \code{\link{read_pepwq}}}

\item{txtsz}{numeric for size of text in the plot, applies only if \code{tab = FALSE}}

\item{trgs}{optional \code{data.frame} for annual bay segment water quality targets, defaults to \code{\link{peptargets}}}

\item{yrrng}{numeric vector indicating min, max years to include}

\item{bay_segment}{chr string for bay segments to include, one to all of "1a", "1b", "2", or "3"}

\item{abbrev}{logical indicating if text labels in the plot are abbreviated as the first letter, applies only to center column}

\item{family}{optional chr string indicating font family for text labels}
}
\value{
A static \code{\link[ggplot2]{ggplot}} object is returned
}
\description{
Create a colorized table for water quality outcomes by segment that includes the management action and chlorophyll, and secchi exceedances
}
\details{
This function provides a combined output for the \code{\link{show_wqmatrixpep}} and \code{\link{show_matrixpep}} functions. Only one bay segment can be plotted for each function call.
}
\examples{
show_segmatrixpep(rawdat, bay_segment = '1a')
}
\seealso{
\code{\link{show_wqmatrixpep}}, \code{\link{show_matrixpep}}
}
\concept{visualize}
