% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_sitemappep.R
\name{show_sitemappep}
\alias{show_sitemappep}
\title{Map water quality data for a selected year}
\usage{
show_sitemappep(
  dat,
  yrsel = NULL,
  mosel = NULL,
  param = c("chla", "sd"),
  bay_segment = c("Western", "Central", "Eastern"),
  maxrel = 0.99,
  relative = FALSE
)
}
\arguments{
\item{dat}{data frame of data returned by \code{\link{read_pepwq}}}

\item{yrsel}{numeric for years to plot, see details}

\item{mosel}{numeric for months to plot, see details}

\item{param}{chr string indicating which water quality value to plot, one of "chla" for chlorophyll and "sd" for secchi depth}

\item{bay_segment}{chr string for the bay segment, one or all of "Western", "Central", or "Eastern"}

\item{maxrel}{numeric for the maximum quantile value for scaling if \code{relative = T}, this prevents outliers from skewing the scale}

\item{relative}{logical indicating if sizes and colors are relative to the entire water quality data base, otherwise scaling is relative only for the points on the map}
}
\value{
A \code{\link[leaflet]{leaflet}} object
}
\description{
Map water quality data for a selected year
}
\details{
Year estimates for the selected parameter are based on median observations across months. All twelve months are used if \code{mosel = NULL} (default). Monthly estimates for the selected parameter are based on median observations across years.  All years are used if \code{yrsel = NULL} (default).

The color ramp is reversed for Secchi observations.
}
\examples{

# 2018, all months
show_sitemappep(rawdat, yrsel = 2018)

# 2018, July only
show_sitemappep(rawdat, yrsel = 2018, mosel = 7)

# July only, all years
show_sitemappep(rawdat, mosel = 7)
}
\concept{visualize}
