% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_thrpep.R
\name{show_thrpep}
\alias{show_thrpep}
\title{Plot annual water quality value and thresholds for a segment}
\usage{
show_thrpep(dat, bay_segment = c("Western", "Central", "Eastern"),
  param = c("chla", "sd"), trgs = NULL, yrrng = c(1976, 2019),
  family = NA, labelexp = TRUE, txtlab = TRUE)
}
\arguments{
\item{dat}{data frame of data returned by \code{\link{read_pepwq}}}

\item{bay_segment}{chr string for the bay segment, one of "Western", "Central", or "Eastern"}

\item{param}{chr string indicating which water quality value and appropriate threshold to plot, one of "chla" for chlorophyll and "sd" for secchi depth}

\item{trgs}{optional \code{data.frame} for annual bay segment water quality thresholds, defaults to \code{\link{peptargets}}}

\item{yrrng}{numeric vector indicating min, max years to include}

\item{family}{optional chr string indicating font family for text labels}

\item{labelexp}{logical indicating if y axis and target labels are plotted as expressions, default \code{TRUE}}

\item{txtlab}{logical indicating if a text label for the target value is shown in the plot}
}
\value{
A \code{\link[ggplot2]{ggplot}} object
}
\description{
Plot annual water quality values and thresholds for a bay segment
}
\examples{
show_thrpep(rawdat, bay_segment = 'Western', param = 'chl')
}
\seealso{
Other visualize: \code{\link{show_boxpep}},
  \code{\link{show_matrixpep}},
  \code{\link{show_reactablepep}},
  \code{\link{show_sitemappep}}
}
\concept{visualize}
