% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_wqmatrixpep.R
\name{show_wqmatrixpep}
\alias{show_wqmatrixpep}
\title{Create a colorized table for chlorophyll or secchi exceedances}
\usage{
show_wqmatrixpep(
  dat,
  param = c("chla", "sd"),
  txtsz = 3,
  trgs = NULL,
  yrrng = NULL,
  alpha = 1,
  bay_segment = c("Western", "Central", "Eastern"),
  asreact = FALSE,
  nrows = 10,
  family = NA
)
}
\arguments{
\item{dat}{\code{data.frame} formatted from \code{\link{read_pepwq}}}

\item{param}{chr string for which parameter to plot, one of \code{"chla"} for chlorophyll or \code{"sd"} for secchi}

\item{txtsz}{numeric for size of text in the plot, applies only if \code{tab = FALSE}}

\item{trgs}{optional \code{data.frame} for annual bay segment water quality targets, defaults to \code{\link{peptargets}}}

\item{yrrng}{numeric vector indicating min, max years to include}

\item{alpha}{numeric indicating color transparency}

\item{bay_segment}{chr string for bay segments to include, one to all of "Western", "Central", or "Eastern"}

\item{asreact}{logical indicating if a \code{\link[reactable]{reactable}} object is returned}

\item{nrows}{if \code{asreact = TRUE}, a numeric specifying number of rows in the table}

\item{family}{optional chr string indicating font family for text labels}
}
\value{
A static \code{\link[ggplot2]{ggplot}} object is returned if \code{asreact = FALSE}, otherwise a \code{\link[reactable]{reactable}} table is returned
}
\description{
Create a colorized table for chlorophyll or secchi exceedances
}
\examples{
show_wqmatrixpep(rawdat)
}
\seealso{
\code{\link{show_matrixpep}}, \code{\link{show_segmatrixpep}}
}
\concept{visualize}
