% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_noaa_license.R
\name{use_noaa_license}
\alias{use_noaa_license}
\title{Document the license for a NOAA package}
\usage{
use_noaa_license(version = 3, include_future = TRUE)
}
\arguments{
\item{version}{An integer specifying which GPL license you wish to use.
Currently, only two or three are allowed and it defaults to three, which
is the latest version as of May 2023.}

\item{include_future}{A logical value. If \code{TRUE}, the result will be to
license your package under the current and any potential future versions
of the license. This is generally considered to be good practice because
it means your package will automatically include "bug" fixes in licenses.
That is, the resulting text will have a great than or equals to sign for
when the license is updated.}
}
\value{
A logical vector indicating if file was modified.
}
\description{
Defines the License portion of your DESCRIPTION file and prints the NOAA
non-license file to the LICENSE, which is ignored in your build but
available for those who wish to view it.
}
\seealso{
\itemize{
\item \code{\link[usethis:licenses]{usethis::use_gpl_license()}}
\item \url{https://choosealicense.com}
\item \href{https://r-pkgs.org/license.html}{license chapter} in \emph{R Packages}.
}
}
\author{
Kelli F. Johnson
}
