% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_other.R
\name{vec_mima_scale}
\alias{vec_mima_scale}
\title{Min-max scale vector}
\usage{
vec_mima_scale(x, na_rm = FALSE)
}
\arguments{
\item{x}{A numeric vector}

\item{na_rm}{Boolean: Should NAs be removed}
}
\description{
Min-max scale vector, such that all values v lie in [0;1]. If x is a numeric
vector, then min-max scale is: ( x - min(x) ) / ( max(x) - min(x) )
}
\examples{
vec_mima_scale(rnorm(20))
}
