% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anlz_dodlypep.R
\name{anlz_dodlypep}
\alias{anlz_dodlypep}
\title{Analyze daily DO values relative to threshold}
\usage{
anlz_dodlypep(dodat, thr = 2.3)
}
\arguments{
\item{dodat}{result returned from \code{\link{read_pepdo}}}

\item{thr}{numeric indicating appropriate dissolved oxygen thresholds, usually 2.3 mg/L for acute, 4.8 mg/L for chronic}
}
\value{
data.frame
}
\description{
Analyze daily DO values relative to threshold
}
\details{
The \code{\link{dodat}} data object can be used as input without downloading USGS data
}
\examples{
data(dodat)
dat <- anlz_dodlypep(dodat)
dat
}
\seealso{
Other analyze: 
\code{\link{anlz_attainpep}()},
\code{\link{anlz_domopep}()},
\code{\link{anlz_entpep}()},
\code{\link{anlz_medpep}()}
}
\concept{analyze}
