% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anlz_dodlypep.R
\name{anlz_dodlypep}
\alias{anlz_dodlypep}
\title{Analyze daily DO values relative to threshold}
\usage{
anlz_dodlypep(dodat, thr = 3, impute = TRUE)
}
\arguments{
\item{dodat}{result returned from \code{\link{read_pepdo}}}

\item{thr}{numeric indicating appropriate dissolved oxygen thresholds, usually 3 mg/L for acute, 4.8 mg/L for chronic}

\item{impute}{logical indicating of missing dissolved oxygen values are imputed with the year, month, site average}
}
\value{
data.frame
}
\description{
Analyze daily DO values relative to threshold
}
\details{
The \code{\link{dodat}} data object can be used as input without downloading USGS data

If \code{impute = TRUE}, missing dissolved oxygen values in the complete daily time series are imputed to the average for the year, month, site combination. This is often necessary to create summary values that make sense. For example, if a month has incomplete data, the maximum \code{below_cumsum} value will not show 30 or 31 days even if every day in the observed record is below the threshold.
}
\examples{
data(dodat)
dat <- anlz_dodlypep(dodat)
dat
}
