% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anlz_domopep.R
\name{anlz_domopep}
\alias{anlz_domopep}
\title{Analyze monthly DO values relative to threshold}
\usage{
anlz_domopep(dodat, thr = 3)
}
\arguments{
\item{dodat}{result returned from \code{\link{read_pepdo}}}

\item{thr}{numeric indicating appropriate dissolved oxygen thresholds, usually 3 mg/L for acute, 4.8 mg/L for chronic}
}
\value{
data.frame
}
\description{
Analyze monthly DO values relative to threshold
}
\details{
The \code{\link{dodat}} data object can be used as input without downloading USGS data
}
\examples{
data(dodat)
dat <- anlz_domopep(dodat)
dat
}
\seealso{
Other analyze: 
\code{\link{anlz_attainpep}()},
\code{\link{anlz_dodlypep}()},
\code{\link{anlz_entpep}()},
\code{\link{anlz_medpep}()}
}
\concept{analyze}
