% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anlz_entpep.R
\name{anlz_entpep}
\alias{anlz_entpep}
\title{Count beach exceedances for enterococcus}
\usage{
anlz_entpep(entdat, thr = 104, cats = c(0, 1, 2))
}
\arguments{
\item{entdat}{result returned from \code{\link{read_pepent}}}

\item{thr}{numeric value defining threshold for exceedance}
}
\value{
A \code{data.frame} with counts of exceedances per year for each beach
}
\description{
Count beach exceedances for enterococcus
}
\details{
The exceedance threshold is set by default as 104 cfu/100 ml criterion.  This is simply based on counts in a year when any value at any station was above the threshold for each 24 hour period in the record.

The \code{samples} column shows how many days of the year were sampled at each beach and the \code{perexceedance} column shows what percentage of samples were above the threshold.
}
\examples{
anlz_entpep(entdat)
}
\seealso{
Other analyze: 
\code{\link{anlz_attainpep}()},
\code{\link{anlz_dodlypep}()},
\code{\link{anlz_domopep}()},
\code{\link{anlz_medpep}()}
}
\concept{analyze}
