% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anlz_medpep.R
\name{anlz_medpep}
\alias{anlz_medpep}
\title{Estimate annual means}
\usage{
anlz_medpep(dat)
}
\arguments{
\item{dat}{\code{data.frame} formatted from \code{\link{read_pepwq}}}
}
\value{
Mean estimates for chlorophyll and secchi
}
\description{
Estimate annual means by segment for chlorophyll and secchi data
}
\examples{
# view average estimates
anlz_medpep(rawdat)
}
\seealso{
Other analyze: 
\code{\link{anlz_attainpep}()},
\code{\link{anlz_entpep}()}
}
\concept{analyze}
