% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{BGFREQS}
\alias{BGFREQS}
\title{Amino Acid Residue Background Frequencies}
\format{
A data frame with 5 rows and 20 columns:
\describe{
\item{name}{One-letter symbol for each of the 20 standard proteogenic amino acids}
\item{Kingdom}{Viruses, Archaea, Bacteria, Eukaryota and All}
}
}
\source{
\url{http://isoelectricpointdb.org/statistics.html}
}
\usage{
BGFREQS
}
\description{
Naturally observed amino acid residue background frequencies are available
from the Proteome-pI DB. Proteome Isoelectric Point Database is a database of
pre-computed isoelectric points for proteomes from different model organisms.
}
\details{
Reference paper:
Proteome-pI: proteome isoelectric point database
Lukasz P. Kozlowski
Nucleic Acids Research, Volume 45, Issue D1, 4 January 2017, Pages D1112–
D1116, https://doi.org/10.1093/nar/gkw978
}
\keyword{datasets}
