% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{BGFREQS}
\alias{BGFREQS}
\title{Amino Acid Residue Background Frequencies}
\format{
A data frame with 5 rows and 21 columns, where rows pertain to Kingdom
and columns to amino acids
}
\source{
\url{https://isoelectricpointdb2.mimuw.edu.pl/statistics.html#aa_stats}
}
\usage{
BGFREQS
}
\description{
Naturally observed amino acid residue background frequencies are available
from the Proteome-pI DB. Proteome Isoelectric Point Database is a database of
pre-computed isoelectric points for proteomes from different model organisms.
}
\details{
Reference paper:
Kozlowski LP. Proteome-pI 2.0: proteome isoelectric point database
update. Nucleic Acids Res. 2022 Jan 7;50(D1):D1535-D1540. doi:
10.1093/nar/gkab944. PMID: 34718696; PMCID: PMC8728302.
}
\keyword{datasets}
