% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{BLOSUM50_pca}
\alias{BLOSUM50_pca}
\title{The BLOSUM50 encoding pca matrix}
\format{
A data frame with with 21 rows and 20 columns
}
\source{
\url{ftp://ftp.ncbi.nih.gov/blast/matrices/BLOSUM50}
}
\usage{
BLOSUM50_pca
}
\description{
The BLOSUM50 matrix transformed for encoding by performing a PCA
and then using n number of PCs for encoding. Below table show the
cumulative explained variation, such that, e.g. incl. 10 PCs, will
correspond to 96\% EV.
}
\details{
\tabular{ll}{
   PC \tab Cumulative EV \cr
   1 \tab 0.375 \cr
   2 \tab 0.565 \cr
   3 \tab 0.661 \cr
   4 \tab 0.744 \cr
   5 \tab 0.802 \cr
   6 \tab 0.852 \cr
   7 \tab 0.89 \cr
   8 \tab 0.918 \cr
   9 \tab 0.945 \cr
   10 \tab 0.96 \cr
   11 \tab 0.973 \cr
   12 \tab 0.981 \cr
   13 \tab 0.987 \cr
   14 \tab 0.991 \cr
   15 \tab 0.995 \cr
   16 \tab 0.997 \cr
   17 \tab 0.999 \cr
   18 \tab 1 \cr
   19 \tab 1 \cr
   20 \tab 1 \cr
}
}
\keyword{datasets}
