% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{BLOSUM62_enc}
\alias{BLOSUM62_enc}
\title{The BLOSUM62 encoding matrix}
\format{
A data frame with with 21 rows and 21 columns
}
\source{
\url{ftp://ftp.ncbi.nih.gov/blast/matrices/BLOSUM62}
}
\usage{
BLOSUM62_enc
}
\description{
The BLOSUM62 matrix transformed for encoding by dividing by 5. Furthermore,
row and column zero vectors representing 'X' has been added.
}
\keyword{datasets}
