% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entdat.R
\docType{data}
\name{entdat}
\alias{entdat}
\title{Raw beach pathogen data from Suffolk County}
\format{
A \code{data.frame} object
}
\usage{
entdat
}
\description{
Raw beach pathogen data from Suffolk County
}
\examples{
\dontrun{
path <- system.file("extdata", "enterodata.xlsx", package="peptools")
entdat <- read_pepent(path)
save(entdat, file = 'data/entdat.RData', compress = 'xz')
}
}
\seealso{
Other utilities: 
\code{\link{beaches}},
\code{\link{dodat}},
\code{\link{pepseg}},
\code{\link{pepstations}},
\code{\link{peptargets}},
\code{\link{rawdat}}
}
\concept{utilities}
\keyword{datasets}
