% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pep_check.R
\name{pep_check}
\alias{pep_check}
\title{Check if a vector of peptides is valid.}
\usage{
pep_check(x, require_length = TRUE)
}
\arguments{
\item{x}{A vector of peptides to be checked}

\item{require_length}{Restrict to also requiring that all peptides have same length}
}
\value{
TRUE if the vector passed the check, otherwise stop() is called
}
\description{
According to criteria:
}
\details{
i.   Is a vector with one or more elements

ii.  Is a character vector

iii. All elements in vector have the same number of characters (depending on the require_length argument)

iv.  Elements only contain allowed amino acid residues \code{'ARNDCQEGHILKMFPSTWYVX-'}
}
\examples{
pep_check(x = c("RQGQDHPTM","RGQKTTDNA","NILYEYWDY"))
pep_check(x = c("RQGQDHPTM","RGQKTTDNA","NILYEYWDYX"), require_length = FALSE)

}
