% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pep_ran.R
\name{pep_ran}
\alias{pep_ran}
\title{Generate Random Peptides}
\usage{
pep_ran(n, k)
}
\arguments{
\item{n}{Number of peptides to generate.}

\item{k}{Length of each peptide.}
}
\value{
A character vector containing the generated random peptides.
}
\description{
This function generates random peptides by sampling amino acid residues.
}
\examples{
pep_ran(n = 10, k = 5)

}
