% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PEPmeeting.R
\name{PEPmeeting}
\alias{PEPmeeting}
\title{Generate Meeting Lead and Scribe one year at a time}
\usage{
PEPmeeting(
  year = format(Sys.time(), "\%Y"),
  people = c("Aaron", "Andi", "Brian", "Chantel", "Jason", "John", "Kelli", "Kiva",
    "Kristin", "Ian", "Melissa", "Vlada")
)
}
\arguments{
\item{year}{The year for which to generate the names. Defaults to the
current year.}

\item{people}{A vector of potential leads.}
}
\description{
Generate Meeting Lead and Scribe one year at a time
}
\examples{
PEPtools::PEPmeeting()
}
