% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peptargets.R
\docType{data}
\name{peptargets}
\alias{peptargets}
\title{Bay segment targets}
\format{
A \code{data.frame} object
}
\usage{
peptargets
}
\description{
Bay segment targets
}
\examples{
\dontrun{
peptargets <- tibble::tibble(
  bay_segment = factor(c('Western', 'Central', 'Eastern'), 
    levels = c('Western', 'Central', 'Eastern')), 
  name = factor(c('Western', 'Central', 'Eastern'), 
    levels = c('Western', 'Central', 'Eastern')), 
  sd_thresh = c(6.5, 6.5, 6.5),
  chla_thresh = c(5.5, 5.5, 5.5)
)
save(peptargets, file = 'data/peptargets.RData', compress= 'xz')
}
}
\seealso{
Other utilities: 
\code{\link{pepstations}},
\code{\link{rawdat}}

Other utilities: 
\code{\link{pepstations}},
\code{\link{rawdat}}
}
\concept{utilities}
\keyword{datasets}
