% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{PEPTIDES}
\alias{PEPTIDES}
\title{An allele data set of random 9-mer peptides}
\format{
A data frame with 132,091 rows and 6 columns
}
\usage{
PEPTIDES
}
\description{
Each peptide has been in silico analysed for binding to six alleles:
\code{HLA-A01:01}, \code{HLA-A02:01}, \code{HLA-A03:01},
\code{HLA-B07:02}, \code{HLA-B08:01}, \code{HLA-B27:05}. Analysis was
performed using netMHCpan4.1b.
}
\keyword{datasets}
