% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rawdat.R
\docType{data}
\name{rawdat}
\alias{rawdat}
\title{Raw data from Suffolk County}
\format{A \code{data.frame} object}
\usage{
rawdat
}
\description{
Raw data from Suffolk County
}
\examples{
\dontrun{
rawdat <- read_xlsx('data/data-raw/Peconics SCDHS WQ data - up to 2019 so far.xlsx', col_types = 'text')
save(rawdat, file = 'data/rawdat.RData', compress = 'xz')
}
}
\concept{utilities}
\keyword{datasets}
