% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rawdat.R
\docType{data}
\name{rawdat}
\alias{rawdat}
\title{Raw data from Suffolk County}
\format{
A \code{data.frame} object
}
\usage{
rawdat
}
\description{
Raw data from Suffolk County
}
\examples{
\dontrun{
path <- system.file("extdata", "currentdata.xlsx", package="peptools")
rawdat <- read_pepwq(path)
save(rawdat, file = 'data/rawdat.RData', compress = 'xz')
}
}
\seealso{
Other utilities: 
\code{\link{beaches}},
\code{\link{dodat}},
\code{\link{entdat}},
\code{\link{pepseg}},
\code{\link{pepstations}},
\code{\link{peptargets}}
}
\concept{utilities}
\keyword{datasets}
