% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_pepdo.R
\name{read_pepdo}
\alias{read_pepdo}
\title{Import dissolved oxygen data}
\usage{
read_pepdo(
  site = c("01304562", "01304200"),
  nms = NULL,
  startDate = "2013-01-01",
  endDate = "2020-12-31"
)
}
\arguments{
\item{site}{chr string of site numbers for USGS stations}

\item{nms}{optional chr string vector of names to reassign to site numbers}

\item{startDate}{chr string of start date in YYYY-MM-DD}

\item{endDate}{chr string of end date in YYYY-MM-DD}
}
\value{
data.frame
}
\description{
Import dissolved oxygen data
}
\details{
Raw data are downloaded using the USGS dataRetrieval R package, this function is a simple wrapper to the \code{\link[dataRetrieval]{readNWISuv}} function.

Note that downloading the station data with the default arguments may take a few minutes. Site are 01304562 for Peconic River, 01304200 for Orient Harbor.
}
\examples{
\dontrun{
dodat <- read_pepdo(site = c('01304562', '01304200'), nms = c('Peconic R.', 'Orient Harbor'), 
     startDate = '2020-06-01', endDate = '2020-06-30')
dodat
}
}
\seealso{
Other read: 
\code{\link{read_pepent}()},
\code{\link{read_pepwq}()}
}
\concept{read}
