% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_pepent.R
\name{read_pepent}
\alias{read_pepent}
\title{Import raw enterococcus data}
\usage{
read_pepent()
}
\value{
data.frame
}
\description{
Import raw enterococcus data
}
\details{
Data are from the ArcGIS REST Services here \url{https://gis.suffolkcountyny.gov/hosted/rest/services/Hosted/Beach_Water_Quality_Data/FeatureServer/}.  

The API is queried by beach names in the \code{\link{beaches}} data object. The queries are done individually for each beach to not exceed the 2000 record limit.
}
\examples{
\dontrun{
entdat <- read_pepent()
}
head(entdat)
}
\concept{read}
