% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_pepent.R
\name{read_pepent}
\alias{read_pepent}
\title{Import raw enterococcus data}
\usage{
read_pepent(path)
}
\arguments{
\item{path}{chr string of path for excel file}
}
\value{
data.frame
}
\description{
Import raw enterococcus data
}
\details{
Raw data from here \url{https://gisportal.suffolkcountyny.gov/gis/home/item.html?id=025cb4dadb57413980dbd7e760b94da8}
}
\examples{
path <- system.file("extdata", "enterodata.xlsx", package="pepreporting")
entdat <- read_pepent(path)
entdat
}
\seealso{
Other read: 
\code{\link{read_pepwq}()}
}
\concept{read}
