% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_boxpep.R
\name{show_boxpep}
\alias{show_boxpep}
\title{Plot monthly chlorophyll or secchi values values for a segment}
\usage{
show_boxpep(dat, param = c("chla", "sd"), yrsel = NULL,
  yrrng = c(1976, 2019), ptsz = 0.5, bay_segment = c("Western",
  "Central", "Eastern"), trgs = NULL, family = NA, labelexp = TRUE,
  txtlab = TRUE)
}
\arguments{
\item{dat}{data frame of data returned by \code{\link{read_pepwq}}}

\item{param}{chr string for which parameter to plot, one of \code{"chla"} for chlorophyll or \code{"sd"} for secchi depth}

\item{yrsel}{numeric for year to emphasize, shown as separate red points on the plot}

\item{yrrng}{numeric vector indicating min, max years to include}

\item{ptsz}{numeric indicating point size of observations not in \code{yrsel}}

\item{bay_segment}{chr string for the bay segment, one of "Western", "Central", or "Eastern"}

\item{trgs}{optional \code{data.frame} for annual bay segment water quality targets, defaults to \code{\link{peptargets}}}

\item{family}{optional chr string indicating font family for text labels}

\item{labelexp}{logical indicating if y axis and target labels are plotted as expressions, default \code{TRUE}}

\item{txtlab}{logical indicating if a text label for the target value is shown in the plot}
}
\value{
A \code{\link[ggplot2]{ggplot}} object
}
\description{
Plot monthly chlorophyll or secchi values values for a bay segment
}
\details{
Points not included in \code{yrsel} are plotted over the box plots using \code{\link[ggplot2]{position_jitter}}. Use \code{ptsz = -1} to suppress.  The dotted line in the plot shows the threshold.
}
\examples{
show_boxpep(rawdat, bay_segment = 'Western')
}
\seealso{
Other visualize: \code{\link{show_matrixpep}},
  \code{\link{show_plotlypep}},
  \code{\link{show_reactablepep}},
  \code{\link{show_sitemappep}}, \code{\link{show_thrpep}},
  \code{\link{show_wqmatrixpep}}
}
\concept{visualize}
