% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_domatrix.R
\name{show_domatrix}
\alias{show_domatrix}
\title{Create a colorized table for reporting dissolved oxygen data by site}
\usage{
show_domatrix(
  dodat,
  site,
  show = c("below_ave", "below_maxrun"),
  txtsz = 3,
  thr = 4.8,
  impute = TRUE,
  yrrng = NULL,
  family = NA
)
}
\arguments{
\item{dodat}{data frame of dissolved oxygen data returned by \code{\link{read_pepdo}}}

\item{site}{character string of the site to plot taken from the \code{nms} argument in \code{\link{read_pepdo}}, usually one of \code{"Peconic River"}, \code{"Orient Harbor"}, or \code{"Shelter Island"}}

\item{show}{chr string indicating which summary value to plot from \code{\link{anlz_domopep}}, one of \code{'below_ave'} or \code{'below_maxrun'}}

\item{txtsz}{numeric for size of text in the plot, applies only if \code{asreact = FALSE}}

\item{thr}{numeric indicating appropriate dissolved oxygen thresholds, usually 3 mg/L for acute, 4.8 mg/L for chronic}

\item{impute}{logical indicating of missing dissolved oxygen values are imputed with the year, month, site average}

\item{yrrng}{numeric vector indicating min, max years to include}

\item{family}{optional chr string indicating font family for text labels}
}
\value{
A static \code{\link[ggplot2]{ggplot}} object is returned.
}
\description{
Create a colorized table for reporting dissolved oxygen data by site
}
\examples{
show_domatrix(dodat, site = 'Peconic River')
}
\seealso{
\code{\link{anlz_domopep}}, \code{\link{anlz_dodlypep}}
}
\concept{visualize}
