% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_entmatrix.R
\name{show_entmatrix}
\alias{show_entmatrix}
\title{Create a colorized table for beach closure reporting}
\usage{
show_entmatrix(
  entdat,
  txtsz = 3,
  thr = 104,
  cats = c(0, 1, 2),
  yrrng = c(2010, 2019),
  asreact = FALSE,
  nrows = 10,
  family = NA
)
}
\arguments{
\item{entdat}{data frame of enterococcus data returned by \code{\link{read_pepent}}}

\item{txtsz}{numeric for size of text in the plot, applies only if \code{asreact = FALSE}}

\item{thr}{numeric value defining threshold for beach closure}

\item{cats}{vector of three numeric values defining the color scheme for the report card}

\item{yrrng}{numeric vector indicating min, max years to include}

\item{asreact}{logical indicating if a \code{\link[reactable]{reactable}} object is returned}

\item{nrows}{if \code{asreact = TRUE}, a numeric specifying number of rows in the table}

\item{family}{optional chr string indicating font family for text labels}
}
\value{
A static \code{\link[ggplot2]{ggplot}} object is returned if \code{asreact = FALSE}, otherwise a \code{\link[reactable]{reactable}} table is returned
}
\description{
Create a colorized table for beach closure reporting
}
\examples{
show_entmatrix(entdat)
}
\seealso{
\code{\link{anlz_entpep}}

Other visualize: 
\code{\link{show_boxpep}()},
\code{\link{show_domatrix}()},
\code{\link{show_matrixpep}()},
\code{\link{show_plotlypep}()},
\code{\link{show_reactablepep}()},
\code{\link{show_segmatrixpep}()},
\code{\link{show_sitemappep}()},
\code{\link{show_thrpep}()},
\code{\link{show_wqmatrixpep}()}
}
\concept{visualize}
