% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_plotlypep.R
\name{show_plotlypep}
\alias{show_plotlypep}
\title{Plot chlorophyll and secchi data together with matrix outcomes}
\usage{
show_plotlypep(
  dat,
  bay_segment = c("Western", "Central", "Eastern"),
  yrrng = NULL
)
}
\arguments{
\item{dat}{\code{data.frame} formatted from \code{\link{read_pepwq}}}

\item{bay_segment}{chr string for the bay segment, one of "Western", "Central", or "Eastern"}

\item{yrrng}{numeric for year range to plot}
}
\value{
An interactive plotly object
}
\description{
Plot chlorophyll and secchi data together with matrix outcomes
}
\examples{
show_plotlypep(rawdat)
}
\seealso{
Other visualize: 
\code{\link{show_boxpep}()},
\code{\link{show_domatrix}()},
\code{\link{show_entmatrix}()},
\code{\link{show_matrixpep}()},
\code{\link{show_reactablepep}()},
\code{\link{show_segmatrixpep}()},
\code{\link{show_sitemappep}()},
\code{\link{show_thrpep}()},
\code{\link{show_wqmatrixpep}()}
}
\concept{visualize}
