% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_sitemappep.R
\name{show_sitemappep}
\alias{show_sitemappep}
\title{Map water quality data for a selected year}
\usage{
show_sitemappep(dat, yrsel, param = c("chla", "sd"),
  bay_segment = c("Western", "Central", "Eastern"))
}
\arguments{
\item{dat}{data frame of data returned by \code{\link{read_pepwq}}}

\item{yrsel}{numeric for year to plot}

\item{param}{chr string indicating which water quality value to plot, one of "chla" for chlorophyll and "sd" for secchi depth}

\item{bay_segment}{chr string for the bay segment, one or all of "Western", "Central", or "Eastern"}
}
\value{
A \code{\link[leaflet]{leaflet}} object
}
\description{
Map water quality data for a selected year
}
\details{
Year estimates for the selected parameter are based on median observations across months.  The color ramp is reversed for Secchi observations.
}
\examples{
show_sitemappep(rawdat, yrsel = 2018)
}
\seealso{
Other visualize: \code{\link{show_boxpep}},
  \code{\link{show_matrixpep}},
  \code{\link{show_plotlypep}},
  \code{\link{show_reactablepep}},
  \code{\link{show_thrpep}}, \code{\link{show_wqmatrixpep}}
}
\concept{visualize}
